Controllers: Refactor CControllerPort into its own file
