+++ xbmc/games/controllers/ControllerTopology.h
@@ -8,7 +8,8 @@
 
 #pragma once
 
-#include <string>
+#include "ControllerPort.h"
+
 #include <vector>
 
 class TiXmlElement;
@@ -17,49 +18,6 @@ namespace KODI
 {
 namespace GAME
 {
-
-class CControllerPort
-{
-public:
-  CControllerPort() = default;
-
-  /*!
-   * \brief Create a controller port
-   *
-   * \param portId The port's ID
-   * \param accepts A list of controller IDs that this port accepts
-   */
-  CControllerPort(std::string portId, std::vector<std::string> accepts);
-
-  void Reset();
-
-  /*!
-   * \brief Get the ID of the port
-   *
-   * \return The port's ID, e.g. "1", as a string
-   */
-  const std::string& ID() const { return m_portId; }
-
-  /*!
-   * \brief Get the controllers that can connect to this port
-   *
-   * \return A list of controllers that are physically compatible with this port
-   */
-  const std::vector<std::string>& Accepts() const { return m_accepts; }
-
-  /*!
-   * \brief Check if the controller is compatible with this port
-   *
-   * \return True if the controller is accepted, false otherwise
-   */
-  bool IsCompatible(const std::string& controllerId) const;
-
-  bool Deserialize(const TiXmlElement* pElement);
-
-private:
-  std::string m_portId;
-  std::vector<std::string> m_accepts;
-};
 
 /*!
  * \brief Represents the physical topology of controller add-ons
